requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pycryptodome==3.19.0
requests==2.31.0
six==1.16.0